requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
airbyte-cdk==7.5.1
airbyte_protocol_models_dataclasses==0.17.1
annotated-types==0.7.0
anthropic==0.70.0
anyascii==0.3.3
anyio==4.11.0
Appium-Python-Client==5.2.4
arrow==1.4.0
asyncer==0.0.9
attributes-doc==0.4.0
attrs==25.4.0
audioop-lts==0.2.2
backoff==2.2.1
bidict==0.23.1
binaryornot==0.4.4
blis==1.3.3
boltons==25.0.0
bracex==2.6
build==1.3.0
CacheControl==0.14.4
cachetools==6.2.2
catalogue==2.0.10
cattrs==25.3.0
certifi==2025.10.5
cffi==2.0.0
chainlit==2.8.3
chardet==5.2.0
charset-normalizer==3.4.4
chevron==0.14.0
cleo==2.1.0
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cookiecutter==2.6.0
crashtest==0.4.1
cryptography==44.0.3
cuid==0.4
cymem==2.0.13
dataclasses-json==0.6.7
dateparser==1.2.2
Deprecated==1.2.18
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
docx==0.2.4
dpath==2.2.0
dulwich==0.24.10
dunamai==1.25.0
et_xmlfile==2.0.0
Faker==38.2.0
fastapi==0.116.2
fastjsonschema==2.21.2
filelock==3.20.0
filetype==1.2.0
findpython==0.7.1
frozenlist==1.8.0
fsspec==2025.9.0
fuzzywuzzy==0.18.0
genson==1.3.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-secret-manager==2.25.0
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
idna==3.11
importlib_metadata==8.7.0
inflection==0.5.1
installer==0.7.0
isodate==0.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==25.7.0
Lazify==0.4.0
Levenshtein==0.27.1
literalai==0.1.201
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mcp==1.17.0
mdurl==0.1.2
monotonic==1.6
more-itertools==10.8.0
msgpack==1.1.2
multidict==6.7.0
murmurhash==1.0.15
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.1
node==1.2.2
numpy==2.3.5
odict==1.9.0
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-alephalpha==0.47.3
opentelemetry-instrumentation-anthropic==0.47.3
opentelemetry-instrumentation-bedrock==0.47.3
opentelemetry-instrumentation-chromadb==0.47.3
opentelemetry-instrumentation-cohere==0.47.3
opentelemetry-instrumentation-crewai==0.47.3
opentelemetry-instrumentation-google-generativeai==0.47.3
opentelemetry-instrumentation-groq==0.47.3
opentelemetry-instrumentation-haystack==0.47.3
opentelemetry-instrumentation-lancedb==0.47.3
opentelemetry-instrumentation-langchain==0.47.3
opentelemetry-instrumentation-llamaindex==0.47.3
opentelemetry-instrumentation-logging==0.59b0
opentelemetry-instrumentation-marqo==0.47.3
opentelemetry-instrumentation-mcp==0.47.3
opentelemetry-instrumentation-milvus==0.47.3
opentelemetry-instrumentation-mistralai==0.47.3
opentelemetry-instrumentation-ollama==0.47.3
opentelemetry-instrumentation-openai==0.47.3
opentelemetry-instrumentation-openai-agents==0.47.3
opentelemetry-instrumentation-pinecone==0.47.3
opentelemetry-instrumentation-qdrant==0.47.3
opentelemetry-instrumentation-redis==0.59b0
opentelemetry-instrumentation-replicate==0.47.3
opentelemetry-instrumentation-requests==0.59b0
opentelemetry-instrumentation-sagemaker==0.47.3
opentelemetry-instrumentation-sqlalchemy==0.59b0
opentelemetry-instrumentation-threading==0.59b0
opentelemetry-instrumentation-together==0.47.3
opentelemetry-instrumentation-transformers==0.47.3
opentelemetry-instrumentation-urllib3==0.59b0
opentelemetry-instrumentation-vertexai==0.47.3
opentelemetry-instrumentation-watsonx==0.47.3
opentelemetry-instrumentation-weaviate==0.47.3
opentelemetry-instrumentation-writer==0.47.3
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.59b0
orjson==3.11.4
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pbs-installer==2025.11.20
pdfminer.six==20250506
phonenumbers==9.0.19
pillow==11.3.0
pkginfo==1.12.1.2
platformdirs==4.5.0
plumber==1.7
poetry==2.2.1
poetry-core==2.2.1
posthog==3.25.0
preshed==3.0.12
presidio_analyzer==2.2.360
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
pypandoc==1.15
pypdf==6.5.0
pyproject_hooks==1.2.0
pyrate-limiter==3.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-engineio==4.12.3
python-Levenshtein==0.27.1
python-multipart==0.0.20
python-slugify==8.0.4
python-socketio==5.14.2
python-ulid==3.1.0
pytz==2024.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
RapidFuzz==3.14.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-cache==1.2.1
requests-file==3.0.1
requests-toolbelt==1.0.0
rich==14.2.0
rich-click==1.9.4
rpds-py==0.27.1
rsa==4.9.1
selenium==4.38.0
serpyco-rs==1.17.1
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
sse-starlette==3.0.2
starlette==0.48.0
syncer==2.0.3
tenacity==9.1.2
text-unidecode==1.3
thinc==8.3.10
tldextract==5.3.0
tokenizers==0.22.1
tomli==2.3.0
tomlkit==0.13.3
tqdm==4.67.1
traceloop-sdk==0.47.3
trio==0.32.0
trio-websocket==0.12.2
trove-classifiers==2025.11.14.15
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
Unidecode==1.4.0
url-normalize==2.2.1
urllib3==2.5.0
uvicorn==0.37.0
virtualenv==20.35.4
wasabi==1.1.3
watchfiles==0.24.0
wcmatch==10.0
weasel==0.4.3
websocket-client==1.9.0
whenever==0.8.10
wrapt==1.17.3
wsproto==1.2.0
xmltodict==0.14.2
yarl==1.22.0
zipp==3.23.0
zope.component==6.1
zope.deferredimport==5.1
zope.deprecation==5.1
zope.event==5.1.1
zope.hookable==7.0
zope.interface==7.2
zope.lifecycleevent==5.1
zope.proxy==6.1
zstandard==0.25.0
